refactor: move tasks to subfolders
