+++ game/tasks/task00_introduction.py
@@ -2,7 +2,12 @@
 
 Welcome to Design Pattern Island!
 
-You will be writing Python code following S.O.L.I.D. principles to help your character escape the island. Each Task teaches one of the principles from SOLID by walking through examples and including a code challenge.
+You will be writing Python code following S.O.L.I.D. principles to help your character
+escape the island. Each Task teaches one of the principles from SOLID by walking through
+examples and including a code challenge.
+
+In each folder, review the introduction file for a background on the principle, then
+make the necessary changes to the task file to advance in the game!
 
 Good luck! To get started, see the Main README.
