+++ game/tasks/task02/task_o_inventory.py
@@ -1,0 +1,19 @@
+"""Task 02: Inventory.
+
+The second task will be to apply the "O" of the S.O.L.I.D design principles to ...
+
+> (O) **Open-Close Principle**
+>
+> Open for extension, not modification.
+
+"""
+"""
+
+
+FYI: Unfortunately we don't yet have a challenge for this task. You will still have the
+opportunity to apply the Open Closed Principle in Task 3, so fret not!
+
+Skip ahead to Task 3 to learn about and apply the Liskov Substitution Principle
+
+
+"""
