+++ game/tasks/task03/lsp_introduction.py
@@ -1,0 +1,26 @@
+"""Task 03: Crafting.
+
+> (L) **Liskov-Substitution Principle**
+>
+> Software entities should be interchangeable.
+
+"""
+"""
+
+
+An important step in designing reusable code is to find opportunities to write code that
+is interchangeable.
+
+
+FYI: We don't have a lesson yet for this task, but we will add one shortly!
+
+"""
+"""
+
+If you would like to learn more about the Liskov-Substitution Principle, you can take a look at:
+
+- https://en.wikipedia.org/wiki/Liskov_substitution_principle
+- https://phoenixnap.com/blog/solid-principles
+- https://ezzeddinabdullah.com/post/solid-principles-lsp-py/
+
+"""
