+++ game/tasks/task03/task_l_crafting.py
@@ -1,7 +1,5 @@
 """Task 03: Crafting.
 
-The third task will be to apply the "L" of the S.O.L.I.D design principles to ...
-
 > (L) **Liskov-Substitution Principle**
 >
 > Software entities should be interchangeable.
@@ -15,21 +13,7 @@ from beartype import beartype
 from beartype.typing import Protocol
 from pydantic import BaseModel
 
-from ..core.models.base_player_inventory import BasePlayerInventory
-
-# ==============================    Part 1 (Learn)    ==============================
-"""
-
-An important step in designing reusable code is to find opportunities to write code that
-is interchangeable.
-
-
-FYI: We don't have a lesson yet for this task, but we will add one shortly!
-
-"""
-
-
-# ==============================    Part 2 (Edit)    ==============================
+from ...core.models.base_player_inventory import BasePlayerInventory
 
 
 @runtime_checkable
@@ -98,14 +82,3 @@ class PlayerInventory(BasePlayerInventory):
             self.inventory[item.name] = item
 
         return self.get_item_one_index(item.name)
-
-
-"""
-
-If you would like to learn more about the Liskov-Substitution Principle, you can take a look at:
-
-- https://en.wikipedia.org/wiki/Liskov_substitution_principle
-- https://phoenixnap.com/blog/solid-principles
-- https://ezzeddinabdullah.com/post/solid-principles-lsp-py/
-
-"""
