+++ game/tasks/task04/isp_introduction.py
@@ -1,0 +1,21 @@
+"""Task 04: The Raft.
+
+> (I) **Interface Segregation Principle**
+>
+> Clients should not be forced to depend upon interfaces that they do not use
+
+"""
+"""
+
+FYI: This introduction hasn't been written yet!
+
+"""
+"""
+
+If you would like to learn more about the Interface Segregation Principle, you can take a look at:
+
+- https://en.wikipedia.org/wiki/Interface_segregation_principle
+- https://phoenixnap.com/blog/solid-principles
+- https://ezzeddinabdullah.com/post/solid-principles-isp-py/
+
+"""
