+++ game/tasks/task04/task_i_the_raft.py
@@ -1,0 +1,18 @@
+"""Task 04: The Raft.
+
+> (I) **Interface Segregation Principle**
+>
+> Clients should not be forced to depend upon interfaces that they do not use
+
+"""
+
+"""
+
+This task hasn't been Implemented yet!
+
+FYI: we are planning to make this task the implementation of raft so that the player
+can escape the small island and make it to the mainland. This task could explore how
+making an abstract GameSprite could break the Interface Segregation Principle if
+methods for both the player and vehicle are expected
+
+"""
