--- game/tasks/task04_i_the_raft.py
@@ -1,34 +1,0 @@
-"""Task 04: The Raft.
-
-The fourth task will be to apply the "I" of the S.O.L.I.D design principles to ...
-
-> (I) **Interface Segregation Principle**
->
-> Clients should not be forced to depend upon interfaces that they do not use
-
-"""
-
-# ==============================    Part 1 (Learn)    ==============================
-
-"""
-
-This task hasn't been Implemented yet!
-
-FYI: we are planning to make this task the implementation of raft so that the player
-can escape the small island and make it to the mainland. This task could explore how
-making an abstract GameSprite could break the Interface Segregation Principle if
-methods for both the player and vehicle are expected
-
-"""
-
-# ==============================    Part 2 (Edit)    ==============================
-
-"""
-
-If you would like to learn more about the Interface Segregation Principle, you can take a look at:
-
-- https://en.wikipedia.org/wiki/Interface_segregation_principle
-- https://phoenixnap.com/blog/solid-principles
-- https://ezzeddinabdullah.com/post/solid-principles-isp-py/
-
-"""
