+++ game/tasks/task05/dip_introduction.py
@@ -1,0 +1,21 @@
+"""Task 05: The New Island.
+
+> (D) **Dependency Inversion Principle**
+>
+> Depend upon abstract interfaces and not concrete implementations
+
+"""
+"""
+
+FYI: This introduction hasn't been written yet!
+
+"""
+"""
+
+If you would like to learn more about the Dependency Inversion Principle, you can take a look at:
+
+- https://en.wikipedia.org/wiki/Dependency_inversion_principle
+- https://phoenixnap.com/blog/solid-principles
+- https://ezzeddinabdullah.com/post/solid-principles-dip-py-copy/
+
+"""
