+++ game/tasks/task05/task_d_the_new_island.py
@@ -1,0 +1,20 @@
+"""Task 05: The New Island.
+
+> (D) **Dependency Inversion Principle**
+>
+> Depend upon abstract interfaces and not concrete implementations
+
+"""
+
+"""
+
+This task hasn't been Implemented yet!
+
+FYI: We are considering making this a task to implement different algorithms for random
+walking vs. follow the player. The task could demonstrate how the implementation
+shouldn't be dependent on the algorithm.
+
+Alternatively, for this task, the user could create logic to generate vegetation and
+static items that can be picked up or used
+
+"""
