--- game/tasks/task05_d_the_new_island.py
@@ -1,36 +1,0 @@
-"""Task 05: The New Island.
-
-The fifth task will be to apply the "D" of the S.O.L.I.D design principles to ...
-
-> (D) **Dependency Inversion Principle**
->
-> Depend upon abstract interfaces and not concrete implementations
-
-"""
-
-# ==============================    Part 1 (Learn)    ==============================
-
-"""
-
-This task hasn't been Implemented yet!
-
-FYI: We are considering making this a task to implement different algorithms for random
-walking vs. follow the player. The task could demonstrate how the implementation
-shouldn't be dependent on the algorithm.
-
-Alternatively, for this task, the user could create logic to generate vegetation and
-static items that can be picked up or used
-
-"""
-
-# ==============================    Part 2 (Edit)    ==============================
-
-"""
-
-If you would like to learn more about the Dependency Inversion Principle, you can take a look at:
-
-- https://en.wikipedia.org/wiki/Dependency_inversion_principle
-- https://phoenixnap.com/blog/solid-principles
-- https://ezzeddinabdullah.com/post/solid-principles-dip-py-copy/
-
-"""
